LPS-193648 refactor AttachmentsUploadResponseHandler.java and TempAttachmentsUploadFileEntryHandler.java
move and create inner class in UploadTempAttachmentMVCActionCommand.java and delete handler class
